// src/app/myModules/reusable/card/interfaces/bubbleProps.interface.ts
export interface BUBBLE {
  content?: string;
  position?: string;
  textAlign?: string;
  lineHeight?: string;
  margin?: string;
  padding?: string;
  width?: string;
  height?: string;
  borderRadius?: string;
  backgroundColor?: string;
  zIndex?: string;
  color?: string;
  display?: string;
}